partially asynchronous snapshot transfer & improve preloading logs

+++ charts/all-in-one/scripts/snapshots/partition/upload_snapshot.sh
@@ -12,8 +12,9 @@ SLACK_WEBHOOK=$2
 CF_DISTRIBUTION_ID=$3
 SNAPSHOT_PATH=$4
 STORE_PATH="$6"
-PRESERVE_PARTITIONS="${7:-false}"
-PART_LENGTH="${8:-3}"
+BYPASS_COPYSTATES="${7:-false}"
+PRESERVE_PARTITIONS="${8:-false}"
+PART_LENGTH="${9:-3}"
 
 function setup_rclone() {
   RCLONE_CONFIG_DIR="/root/.config/rclone"
@@ -77,7 +78,7 @@ function make_and_upload_snapshot() {
     echo "URL does not exist: $URL"
   fi
 
-  if ! "$SNAPSHOT" --output-directory "$OUTPUT_DIR" --store-path "$STORE_PATH" --block-before 0 --apv "$APP_PROTOCOL_VERSION" --snapshot-type "partition"; then
+  if ! "$SNAPSHOT" --output-directory "$OUTPUT_DIR" --store-path "$STORE_PATH" --block-before 0 --apv "$APP_PROTOCOL_VERSION" --snapshot-type "partition" --bypass-copystates="$BYPASS_COPYSTATES"; then
     senderr "Snapshot creation failed."
     exit 1
   fi
@@ -112,25 +113,27 @@ function make_and_upload_snapshot() {
     --s3-disable-checksum \
     --s3-copy-cutoff 1G \
     --retries 5 \
-    --low-level-retries 10
+    --low-level-retries 10 &
   retry_until_success rclone copyto "$ARCHIVED_SNAPSHOT_PATH" "$DEST_PATH/latest.zip" \
     --no-traverse \
     --s3-disable-checksum \
     --s3-copy-cutoff 1G \
     --retries 5 \
-    --low-level-retries 10
+    --low-level-retries 10 &
   retry_until_success rclone copyto "$ARCHIVED_SNAPSHOT_PATH" "$DEST_PATH/internal/$SNAPSHOT_FILENAME" \
     --no-traverse \
     --s3-disable-checksum \
     --s3-copy-cutoff 1G \
     --retries 5 \
-    --low-level-retries 10
+    --low-level-retries 10 &
   retry_until_success rclone copyto "$ARCHIVED_SNAPSHOT_PATH" "$DEST_PATH/internal/latest.zip" \
     --no-traverse \
     --s3-disable-checksum \
     --s3-copy-cutoff 1G \
     --retries 5 \
-    --low-level-retries 10
+    --low-level-retries 10 &
+
+  wait
 
   if [ -n "$LATEST_METADATA" ]; then
     echo "[INFO] Archiving metadata..."
@@ -138,11 +141,13 @@ function make_and_upload_snapshot() {
     rclone copyto "$LATEST_METADATA" "$ARCHIVED_METADATA_PATH" --no-traverse --retries 5 --low-level-retries 10
 
     echo "[INFO] Copying metadata to latest path..."
-    rclone copyto "$ARCHIVED_METADATA_PATH" "$DEST_PATH/$METADATA_FILENAME" --no-traverse --retries 5 --low-level-retries 10
-    rclone copyto "$ARCHIVED_METADATA_PATH" "$DEST_PATH/latest.json" --no-traverse --retries 5 --low-level-retries 10
-    rclone copyto "$ARCHIVED_METADATA_PATH" "$DEST_PATH/internal/$METADATA_FILENAME" --no-traverse --retries 5 --low-level-retries 10
-    rclone copyto "$ARCHIVED_METADATA_PATH" "$DEST_PATH/internal/latest.json" --no-traverse --retries 5 --low-level-retries 10
-    rclone copyto "$ARCHIVED_METADATA_PATH" "$DEST_PATH/internal/mainnet_latest.json" --no-traverse --retries 5 --low-level-retries 10
+    rclone copyto "$ARCHIVED_METADATA_PATH" "$DEST_PATH/$METADATA_FILENAME" --no-traverse --retries 5 --low-level-retries 10 &
+    rclone copyto "$ARCHIVED_METADATA_PATH" "$DEST_PATH/latest.json" --no-traverse --retries 5 --low-level-retries 10 &
+    rclone copyto "$ARCHIVED_METADATA_PATH" "$DEST_PATH/internal/$METADATA_FILENAME" --no-traverse --retries 5 --low-level-retries 10 &
+    rclone copyto "$ARCHIVED_METADATA_PATH" "$DEST_PATH/internal/latest.json" --no-traverse --retries 5 --low-level-retries 10 &
+    rclone copyto "$ARCHIVED_METADATA_PATH" "$DEST_PATH/internal/mainnet_latest.json" --no-traverse --retries 5 --low-level-retries 10 &
+
+    wait
   fi
 
   echo "[INFO] Archiving state..."
@@ -162,14 +167,16 @@ function make_and_upload_snapshot() {
     --s3-disable-checksum \
     --s3-copy-cutoff 1G \
     --retries 5 \
-    --low-level-retries 10
+    --low-level-retries 10 &
 
   retry_until_success rclone copyto "$ARCHIVED_STATE_PATH" "$DEST_PATH/internal/$STATE_FILENAME" \
     --no-traverse \
     --s3-disable-checksum \
     --s3-copy-cutoff 1G \
     --retries 5 \
-    --low-level-retries 10
+    --low-level-retries 10 &
+
+  wait
 
   for file in $( find "$OUTPUT_DIR" -size +4G ); do
     split -b 4GB "$file" "$file.part" --numeric-suffixes=1 -a $PART_LENGTH
@@ -181,10 +188,11 @@ function make_and_upload_snapshot() {
         --multi-thread-streams 4 \
         --retries 5 \
         --low-level-retries 10 \
-        --no-traverse
-      rm "$part"
+        --no-traverse && rm "$part" &
     done
   done
+
+  wait
 
   if [ "$PRESERVE_PARTITIONS" = "false" ]; then
     rm "$LATEST_SNAPSHOT" "$LATEST_STATE" "$LATEST_METADATA"
